Switched over to use Spring Boot BasicErrorController instead of our custom ExceptionHandlerController.
